frdp-display, session: Handle mouse events
This opens space for other event handlers to be introduced.

Gtk/Gdk events on FrdpDisplay are forward to FrdpSession which
maps them to FreeRDP events.

Events are processed as they come.

Alternatively we could start pushing events into a queue and
process them in FrdpSession.update. Not sure if this is necessary
or even more efficient. This is what Vinagre does.