frdp-session: Fix livelock when certificate is changed
Currently, gtk-frdp loops forever when the certificate is changed.
Let's silently accept the changed certificate to fix the loop.

It is bad to accept certificates without informing the user, but
I suppose that this change is acceptable given the fact that the
inital certificate is also silently used... before gtk-frdp will
get proper API to handle certificates.

Fixes: https://gitlab.gnome.org/GNOME/gtk-frdp/-/issues/21