clipboard: Do not copy pasted files twice
Previously, files pasted to local filesystem were stored in a temporary
directory from which GtkCLipboard copied them to their final destination.

This commit implements the list of files on remote system as a virtual
filesystem using FUSE. This allows us to give GtkClipboard just list of URIs
on local system from which it copies the files.

This has been inspired by Pascal Nowack's implementation of the same approach
on server side in gnome-remote-desktop project.