clipboard: Do not process received data after finalize
Since receiving some clipboard data can take some time, they could
be received after finalize of FrdpChannelClipboard. This could
lead to crash or data corruption because gtk3's gtk_clipboard_request_contents()
does not allow to use a GCancellable for such situation. So we need to check
"manually" whether the FrdpChannelClipboard was already finalized
when the data arrive. Since the GtkClipboard is the only object
which we know was not finalized I've implemented storing of info
about finalized FrdpChannelClipboards inside of the object.
This happens only when there is a request which was not
fulfilled yet in the time of finalize of FrdpChannelClipboard.