clipboard: Retain file lists with clipDataId
Store file list when a lock request is issued. Store the file list
with its clipDataId so that we know which files are requested
later.
Since we don't know the list of files ourselves right after sending
client format list we have to wait until the data are requested
and obtained from local GtkClipboard to store the list of files.

Request locking of files on the remote side when requesting
files from it.

Fixes #44