gtk4: Fix drawing, compilation and key events
There are still things to finish like scrolling and clipboard.

Unfortunately, clipboard works differently with available data
currently. It seems that it is not possible to pass just formats
to GdkClipboard, the data has to be passed at the same time.

Scrolling has to be reworked in relation to MR !144 which improves
it.