frdp-session: Turn off asynchronous connection
Turn off the asynchronous connection via GTask thread as the FreeRDP
process runs in the new thread which makes it hard to cooperate with
the original thread running UI leading to random race conditions.
Turn it on again if there will be an async support for connection
added to FreeRDP.

The disadvantage is that the application freezes during connection
for some time.