frdp-session: Queue draw area from main thread
frdp_end_paint() can be called from different threads. Since it
queues drawing of areas on the display it should be called from
the main thread only.

This commit adds queue for such requests. The queue is then executed
in the update() function which runs in the main thread.

This should avoid some of the crashes which we see in
https://retrace.fedoraproject.org/faf/problems/.